VERSION UPDATE
Updated plot_hisafe_scene() with improvement to HISAFE where the "overflow" of intercrop across toric symmetry only occurs when both toric symmetry parameters in that direction are activated.

Added line to read_hisafe() so that if a hip is passed and profiles is left to the default ("all"), only the profiles from that hip are attempted to read, rather than all possible profiles.

Updated names of the "default" templates in extdata, and then updated their names as argument defaults in several functions.

Removed waterToDesaturation from all water balance calcs, as we determined that it was irrelevant.

Generalized in build_structure() the overwriting of template PLD, SIM, and WTH files so that they don't necesarily have to be named "template.SIM" etc. within the template directories. This will be critical when users are using their own template directories rather than the built in templates.

Removed from build_structure() the parameters that were confirmed as irrelevant and no longer included in the PLD file.