VERSION UPDATE
Added hisafe_slice(), hisafe_snapshot(), and visual_legend(), with major modifications to plot_hisafe_cells(). This trio is now working very well to the extent of exporting daily "snapshoot" plots. Currently, however, no function is supported to combine exported images into an animation/video. In addition, the legend is also exported as a separate file (done partly to reduce the size/creation time of the primary files and partly because I could not get the legend to align properly with the main plots).

The major modification to plot_hisafe_cells() is that the plotted variable is now scaled to to the range of values in rel.dates. This mimicks the approach used by hisafe_slice() and provides better contrast to long-term animations/videos. Also added a for.anim parameter that toggles on a change in plot formatting. I don't like this approach, but it was the best I could think of for now.

Modified diag_output_path() so it could also be used by hisafe_snapshot().

Modified hop_filter() so that it doesn't check annual and month profiles for presence of dates.

Modified plot_hisafe_cells(), plot_hisafe_monthcells(), and plot_hisafe_annualcrop() to use geom_raster() rather than geom_tile(). This is supposed to be faster, but, more importantly, by using hjust = 1 and vjust = 1, I was able to get the axis labels where they needed to be (i.e. not in the center of cells). As a result of this change, I no longer needed to shift trees by 1/2 a cellWidth before plotting.

Added a small check to fix a bug in read_hisafe() that occured when a requested simu.name is not present in the summary file.

Added a ... argument to both hisafer themes to make them more useful when applying on their own.

Fixes #73
Fixes #74