Added argument plot.scene to build_hisafe() so that the (lengthy) process of building scene maps is an option.
Fixed a bug in plot_hisafe_cycle_annual() in which I had intended to remove incomplete simulation years from being plotted. I was originally removing ALL data from these years, but it is not fixed so that only incomplete Simulation-Years are removed. This allows complete years from Simulations which run longer to remain plotted.

Fixed bug in hop_filter() so that profile_check() is not called because profile_check() only works with output data profiles, not others that are created in the process of read_hisafe().

Fixed a few miscellaneous bugs in hisafe_slice().

Fixed documentation bugs in hisafe_snapshot().

Added a rel.dates argument to hisafe_snapshot() so that a separate set of rel.dates could be maintained even if you only wanted to plot a few dates of a simulation.