FIxed a bug in plot_hisafe_monthcells() where tree canopy data was not being faceted correctly.
Fixed a bug in plot_hisafe_cells() where the Y.MAX was incorrectly the same as X.MAX.

Added phantom tree (wrapping around toric symmetry) functionality to both hisafe_slice() and plot_hisafe_cells() (only when for.anim = TRUE).

Increased spacing between facets in both hisafe_slice() and plot_hisafe_cells().

Tidied up some code in hisafe_slice().

Removed option to turn off the various components of hte legend in visual_legend(). This made it simpler and removes the issues that arise in the lcoation of the scale_fill bar if you turn of any of the other components.