Added support in hisafe_slice() and plot_hisafe_cells() for "memory" of prior cells/voxels data when export frequency is not daily (via helper function add_historic_data()).
Fixed bug in plot_hisafe_cycle_daily() and plot_hisafe_cycle_annual()where error was occuring when trying to plot carbon cycle for a simulation without any trees. Check is now added to simply return NULL if the hop does not have a trees profile.

Fixed bug in create_tree_data() when it was called for a simulation without any trees.

Updated diag_hisafe_cells() with smaller default for dates argument so that plots were of reasonable size. Kept this change within the argument so that a diagnostic run could still be make with customized specific dates.

Removed the dates check in hop_filter() because it fails as soon as daily outputs are exported any less frequent than 1 day. This check is really impossible to make because of the unknown export frequency.