Modified hisafe_slice() so that voxels (along with all other vars now) have aesthetics relative to simulation-year combos. This makes the visualization dramatically better so that roots are easier to see earlier on in the simulation.
Added a new argument (bulk.pass) to define_hisafe() so that a list of hisafe params could be passed, rather than just listing them out directly as arguments. This helps to pass the same set of arguments to multiple calls of define_hisafe(), as I do for the hisafe calibration process.