VERSION UPDATE
(first changes since first release of hisafer to lab)

hisafe_slice() no longer throw and error if crops/voxels = TRUE and it cannot find the cells/voxels profiles. Instead it just switches the logicals to FALSE and does not plot them. This makes the function more amenable to programming and complex situations.