Enhanced MAX.Y determination in hisafe_slice() to be dependent on if trees and crops are plotted or not and to not assume that tree trees will be taller than the crops.
Added a few more forgotten variable names to the list of required ones by hisafe_slice().

Fixed bug in plot_hisafe_monthcells() where the non-faceted variable was not being correctly checked for length > 1. This was resulting in plots that had multiple tree canopies when only a single value for one of the facets was used.

Added all parallel computing packages to the dependencies list and removed the package check from run_hisafe() because this was causing install problems. Also added egg to the dependencies list.