VERSION UPDATE
Set package settings to not build vignettes when building source or binary packages because I always rebuild the vignettes directly after editing them anyways.

Fixed a bug that only appeared when testing hisafer on Windows. The simualtion directory was being created both by build_hisafe() directly and then AGAIN by build_structure() when copying over the template directory. This worked fine on OSx, but on Windows it wouldn't copy the template directory because an (empty) directory with the destination name already existed. Simple fix is to remove the redundant directory creation in build_hisafe().