VERSION UPDATE
Added functionality to plot_hisafe_ts() so that multiple variables can be plotted at once. This required cowplot to be added as a dependent package.

Fixes #78

Made climate (precipitation) also optional in hisafe_slice() so that it can be ignored if the climate profile was not exported.

Discovered base::basename() and substituted that in for the crazy tail(strsplit()) approach I had been using to do the same thing throughout hisafer.