VERSION UPDATE
Added functionality to build_hisafe() so that it will only "build" the specified files, via the argument "files". In reality, build_hisafe() builds ALL files, and then just deletes the folders/files that are not desired. Consequently, specifying not to build certain files does not increase the speed of build_hisafe(), but this functionality is nevertheless useful when some files are not required (e.g. in some cluster situations).

Also added descriptions of each profile to the hisafe_profiles() call, and removed the non-"public" profiles from what is displayed.