Updated cycles.R so that stemflow is now not consider in the water balance because I discovered in the model that stem flow is just REPOSITIONED rain (to the cell where the tree is rooted). It is not lost.
Added a few more hip checks regarding STICS parameter interactions.

Improved the reading of "Detail" profiles so that for trees, plot, cells, and voxels now ANY profile name can be accepted as long as it begins with the proper roots. This is good prep for any future flexibility with profile names.

Added "Detail" profiles to SUPPORTED.PROFILES for plot, trees, cells, and voxels.

Updated hisafe_slice() with new output variable names for grainBiomass and fertilisation.