VERSION UPDATE
A range of updates for the purpose of moving as many packages as possilbe from Imports to Suggests. Was successful in doing this for most packages since many of them are only use for 1-2 function within a single hisafer function. Warnings or errors are now displayed if the required library is not available when it is needed. Most tidyverse packages are still required, as well as R.utils.

Also enhanced functionalit of read_hisafe() so that the generality of profile names applies to all types except climate.

Fixes #96.