Added functionality so that when NA is supplied to any parameter via define_hisafe() it is commented out in the restulting hisafe parameter file. This is very useful for turning "off" things like tree pruning.
Also, fixed bug with template.subpath.

Also, made more flexible how multiple values in the "accepted" column are separated in the INPUT.DEFS.

Unlink the link to cowplot in the documentation of plot_hisafe_ts() because this package is no longer a dependency, so if it is not installed when hisafer is installed this documentaiton link produces a warning.