VERSION UPDATE
Fixed bug in plot_hisafe_scene() when plotting monocrop. This changed because of the addition recently made to plot id numbers on the trees.

Also, base::normalizePath() is suddenly not working when the directy has not yet been created. The documentation says it has unpredictable results in this case. Consquently, I've switched the approach of get_absolute_path() to be a homemade solution. Seems simple enough. I'm sure base::normalizePath() or R.utils::getAbsolutePath can handle many other situations, but the only situation I really need this to handle is "./".